chore: deprecate nlu_methods.js
