+++ botfront/imports/ui/components/nlu/import-export/DeleteModel.jsx
@@ -7,7 +7,7 @@ import 'brace/mode/json';
 import 'brace/theme/github';
 import { saveAs } from 'file-saver';
 import moment from 'moment';
-import { getTrainingDataInRasaFormat } from '../../../../lib/nlu_methods';
+import { getTrainingDataInRasaFormat } from '../../../../api/instances/instances.methods';
 
 export default class DeleteModel extends React.Component {
     constructor(props) {
